requirements.txt
flask
flask-cors
boto3
requests
bcrypt
pyjwt
